Add unit test example for using config xml fixture
Added a working unit test to show how to set config xml via a fixture
and how to read the config xml using assert statements.

Note that the confix_xml specified in the fixture 'extends' an existing
config node, therefore that node must already exist in the config. The
only way to do that is to make sure that you also add the parent node in
config.xml when needed.