Rename config fixture file
Rename Config/_data/fx-config.yaml to fx-config_config.yaml as the
fixture loader borks when there are two fixture files with the same name
in a different position in the fixture loader hierarchy.

I was trying to add another file called Test/_data/fx-config.yaml, but
it broke the existing tests. The solution seems to be to add a prefix to
the config.yaml file that inclusdes the name of the test class that it
relates to.