Clear cacheObserver config element before test
The test was failing when integrated with a project that had config
values set for real cacheObservers. This change was only needed to
ensure the assertion for checking that the config node matched the
expectation would pass.

It is not needed for the other tests, as they are not checking the
config directly, and the test configs can safely be merged with 'real'
configs as they use _test namespaces under the aligent_cacheobserver
namespace.